split of compilerservice from main server
