+++ src/inmanta/server/__init__.py
@@ -23,3 +23,4 @@ SLICE_AGENT_MANAGER = "agentmanager"
 SLICE_SESSION_MANAGER = "session"
 SLICE_DATABASE = "core.database"
 SLICE_TRANSPORT = "core.transport"
+SLICE_COMPILER = "core.compiler"
